refactor: hex formatting and byte handling leave utils.js
hexbyte, hexword, the hex dump and parseAddr move to hex.js; signExtend,
the string and integer readers, DataStream, makeFast32 and the Fifo move to
binary.js. Two files that aliased utils.signExtend and friends into local
consts import them directly instead.

Part of #1072.
