+++ src/disc.js
@@ -1,8 +1,8 @@
 // Translated from beebjit by Chris Evans.
 // https://github.com/scarybeasts/beebjit
 
-import { hexbyte } from "./utils.js";
 import { crc32 } from "./archive.js";
+import { hexbyte } from "./hex.js";
 class TrackBuilder {
     /**
      * @param {Track} track
