+++ src/teletext.js
@@ -1,6 +1,6 @@
 import { makeChars } from "./teletext_data.js";
-import { makeFast32 } from "./utils.js";
 import { BbcDefaultPalette as BbcDefaultCollook } from "./bbc-palette.js";
+import { makeFast32 } from "./binary.js";
 
 export class Teletext {
     constructor() {
