+++ src/web/debug.js
@@ -1,5 +1,6 @@
-import { hexbyte, hexword, noop, parseAddr } from "../utils.js";
+import { noop } from "../utils.js";
 import { toggle } from "../dom-utils.js";
+import { hexbyte, hexword, parseAddr } from "../hex.js";
 
 const numToShow = 16;
